+++ changelog.md
@@ -1,3 +1,5 @@
+Version v2.2.2
+
 - Custom AI: Number of improvements to collision avoidance logic. Ships are now less likely to bump into allies.
 - Custom AI: Ships now avoid mines.
 - Fixed log spam when calculations result in Not a Number. Suggested by Oddish.